extract CounterGather class into its own file
